// src/components/ui/data-display/data-display.types.ts
export type BoxAS = "div" | "span" | "aside" | "main" | "header" | "section" | "footer" | "nav";

export type TModalSizes = "xs" | "sm" | "md" | "lg" | "xl" | "xxl" | "full";
export type TModalPlacement = "top" | "bottom" | "center" | "left" | "right";

export type TextAs = "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "span" | "small" | "p" | "label";
// text
export type TextVariant =
  | "h1"
  | "h2"
  | "h3"
  | "h4"
  | "h5"
  | "h6"
  | "display1"
  | "display2"
  | "display3"
  | "small"
  | "overline"
  | "paragraph"
  | "label";

export type TextAlign = "center" | "inherit" | "justify" | "left" | "right" | "start" | "end";

// Avatar
export type AvatarVariant = "image" | "initials" | "icon";
export type AvatarSubVariant = "normal" | "status" | "icon";
// Badge
export type BadgeVariant = "gray" | "info" | "success" | "warning" | "primary" | "danger";
// Note
export type NoteVariant = "primary" | "default" | "warning" | "success" | "danger";
// Tab
export type TabVariant = "default" | "shadow" | "bottom-border" | "bottom-border-bg";